# openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "auth.ericminassian.com",
    "description": "Personal OIDC provider — SPA API and token endpoint.",
    "license": {
      "name": "MIT"
    },
    "version": "0.1.0"
  },
  "paths": {
    "/api/account": {
      "delete": {
        "tags": [
          "account"
        ],
        "summary": "DELETE /api/account — permanently delete the account and everything bound\nto it: passkeys, sessions (each cascading to refresh families +\nback-channel logout), then the user record itself.",
        "operationId": "delete_account",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/account/passkeys": {
      "get": {
        "tags": [
          "account"
        ],
        "summary": "GET /api/account/passkeys",
        "operationId": "list_passkeys",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PasskeyList"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/account/passkeys/{credential_id}": {
      "delete": {
        "tags": [
          "account"
        ],
        "summary": "DELETE /api/account/passkeys/{credential_id} — refuses to delete the last\npasskey (account lockout guard; recovery would be the only way back in).",
        "operationId": "delete_passkey",
        "parameters": [
          {
            "name": "credential_id",
            "in": "path",
            "description": "base64url credential id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Cannot delete the only passkey",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "account"
        ],
        "summary": "PATCH /api/account/passkeys/{credential_id}",
        "operationId": "rename_passkey",
        "parameters": [
          {
            "name": "credential_id",
            "in": "path",
            "description": "base64url credential id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RenameRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/account/sessions": {
      "get": {
        "tags": [
          "account"
        ],
        "summary": "GET /api/account/sessions",
        "operationId": "list_sessions",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionList"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/account/sessions/{session_id}": {
      "delete": {
        "tags": [
          "account"
        ],
        "summary": "DELETE /api/account/sessions/{session_id} — session_id is the sid hash as\nreturned by the list endpoint (opaque to the client).",
        "operationId": "revoke_session",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "description": "Opaque session id from the list endpoint",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/recovery/start": {
      "post": {
        "tags": [
          "recovery"
        ],
        "summary": "POST /api/recovery/start — sends a recovery OTP if the account exists.\nUniform 200 either way (anti-enumeration).",
        "operationId": "start",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/recovery/verify": {
      "post": {
        "tags": [
          "recovery"
        ],
        "summary": "POST /api/recovery/verify — consumes the OTP and issues an enroll-level\nsession on the existing account so a new passkey can be registered.",
        "operationId": "verify",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Enroll-level session cookie set",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerifyResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid or expired code",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "description": "Error envelope returned by every `/api/*` failure.",
        "required": [
          "error",
          "message"
        ],
        "properties": {
          "error": {
            "type": "string",
            "description": "Machine-readable error code, e.g. `rate_limited`, `account_exists`."
          },
          "message": {
            "type": "string"
          }
        }
      },
...
      "OkResponse": {
        "type": "object",
        "description": "Generic success envelope (`{ \"ok\": true }`).",
        "required": [
          "ok"
        ],
        "properties": {
          "ok": {
            "type": "boolean"
          }
        }
      },
      "PasskeyInfo": {
        "type": "object",
        "required": [
          "credential_id",
          "name",
          "created_at"
        ],
        "properties": {
          "created_at": {
            "type": "integer",
            "format": "int64"
          },
          "credential_id": {
            "type": "string"
          },
          "last_used_at": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "name": {
            "type": "string"
          }
        }
      },
      "PasskeyList": {
        "type": "object",
        "required": [
          "passkeys"
        ],
        "properties": {
          "passkeys": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PasskeyInfo"
            }
          }
        }
      },
...
      "RenameRequest": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string"
          }
        }
      },
...
      "SessionList": {
        "type": "object",
        "required": [
          "sessions"
        ],
        "properties": {
          "sessions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SessionListItem"
            }
          }
        }
      },
      "SessionListItem": {
        "type": "object",
        "required": [
          "session_id",
          "created_at",
          "last_seen_at",
          "amr",
          "current"
        ],
        "properties": {
          "amr": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "created_at": {
            "type": "integer",
            "format": "int64"
          },
          "current": {
            "type": "boolean"
          },
          "last_seen_at": {
            "type": "integer",
            "format": "int64"
          },
          "session_id": {
            "type": "string"
          }
        }
      },
...
      "StartRequest": {
        "type": "object",
        "required": [
          "email"
        ],
        "properties": {
          "email": {
            "type": "string"
          }
        }
      },
...
      "VerifyRequest": {
        "type": "object",
        "required": [
          "email",
          "code"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "email": {
            "type": "string"
          }
        }
      },
      "VerifyResponse": {
        "type": "object",
        "required": [
          "user_id"
        ],
        "properties": {
          "user_id": {
            "type": "string",
            "format": "uuid"
          }
        }
      }
    }
  },
...
}